feat: refactor cookie consent handling into shared test helpers for improved code reuse
