+++ e2e/time-tracker-privacy.spec.js
@@ -1,29 +1,12 @@
 import { test, expect } from "@playwright/test";
-
-// Helper function to pre-set cookie consent in localStorage (prevents dialog from appearing)
-async function presetCookieConsent(page, url = "/") {
-  // First navigation: Load page to get access to localStorage
-  await page.goto(url);
-  await page.evaluate(() => {
-    localStorage.setItem(
-      "cookie-consent-preferences",
-      JSON.stringify({
-        version: "1.0",
-        necessary: true,
-        analytics: false,
-        timestamp: new Date().toISOString(),
-      })
-    );
-  });
-  // Second navigation: Reload page with consent already set
-  // This ensures the page's cookie consent logic sees the preference on load
-  // and skips showing the dialog entirely
-  await page.goto(url);
-}
+import { handleCookieConsent } from "./test-helpers.js";
 
 test.describe("Time Tracker Privacy Policy Page", () => {
   test("should load privacy policy page successfully", async ({ page }) => {
-    await presetCookieConsent(page, "/time-tracker-privacy.html");
+    await page.goto("/time-tracker-privacy.html");
+
+    // Handle cookie consent
+    await handleCookieConsent(page);
 
     // Check page title
     await expect(page).toHaveTitle(/Time Tracker.*Privacy Policy/);
@@ -36,7 +19,10 @@ test.describe("Time Tracker Privacy Policy Page", () => {
   });
 
   test("should have proper privacy policy structure", async ({ page }) => {
-    await presetCookieConsent(page, "/time-tracker-privacy.html");
+    await page.goto("/time-tracker-privacy.html");
+
+    // Handle cookie consent
+    await handleCookieConsent(page);
 
     // Check for key sections
     await expect(
@@ -58,7 +44,10 @@ test.describe("Time Tracker Privacy Policy Page", () => {
   });
 
   test("should display effective date", async ({ page }) => {
-    await presetCookieConsent(page, "/time-tracker-privacy.html");
+    await page.goto("/time-tracker-privacy.html");
+
+    // Handle cookie consent
+    await handleCookieConsent(page);
 
     // Check for date information
     const content = await page.textContent("body");
@@ -67,7 +56,10 @@ test.describe("Time Tracker Privacy Policy Page", () => {
   });
 
   test("should have working navigation back to main site", async ({ page }) => {
-    await presetCookieConsent(page, "/time-tracker-privacy.html");
+    await page.goto("/time-tracker-privacy.html");
+
+    // Handle cookie consent
+    await handleCookieConsent(page);
 
     // Check navigation links exist (use first to avoid strict mode violation)
     const homeLink = page.locator('a[href="/#home"]').first();
@@ -84,7 +76,10 @@ test.describe("Time Tracker Privacy Policy Page", () => {
   test("should have consistent theme toggle functionality", async ({
     page,
   }) => {
-    await presetCookieConsent(page, "/time-tracker-privacy.html");
+    await page.goto("/time-tracker-privacy.html");
+
+    // Handle cookie consent
+    await handleCookieConsent(page);
 
     const themeToggle = page.locator("#theme-toggle");
     const html = page.locator("html");
@@ -105,8 +100,10 @@ test.describe("Time Tracker Privacy Policy Page", () => {
   });
 
   test("should persist theme across navigation", async ({ page }) => {
-    // Pre-set cookie consent to prevent dialog from appearing
-    await presetCookieConsent(page, "/");
+    await page.goto("/");
+
+    // Handle cookie consent
+    await handleCookieConsent(page);
 
     const html = page.locator("html");
     const themeToggle = page.locator("#theme-toggle");
@@ -128,7 +125,7 @@ test.describe("Time Tracker Privacy Policy Page", () => {
     // Theme should still be dark
     await expect(html).toHaveAttribute("data-theme", "dark");
 
-    // Navigate back to main page (consent is already in localStorage)
+    // Navigate back to main page
     await page.goto("/");
 
     // Theme should still be dark
@@ -136,7 +133,10 @@ test.describe("Time Tracker Privacy Policy Page", () => {
   });
 
   test("should have accessible contact information", async ({ page }) => {
-    await presetCookieConsent(page, "/time-tracker-privacy.html");
+    await page.goto("/time-tracker-privacy.html");
+
+    // Handle cookie consent
+    await handleCookieConsent(page);
 
     // Check for contact email (use first to avoid strict mode violation)
     const emailLink = page.locator('a[href*="mailto"]').first();
@@ -150,7 +150,10 @@ test.describe("Time Tracker Privacy Policy Page", () => {
   test("should have links to third-party privacy policies", async ({
     page,
   }) => {
-    await presetCookieConsent(page, "/time-tracker-privacy.html");
+    await page.goto("/time-tracker-privacy.html");
+
+    // Handle cookie consent
+    await handleCookieConsent(page);
 
     // Check for external policy links
     const content = await page.textContent("body");
@@ -173,7 +176,10 @@ test.describe("Time Tracker Privacy Policy Page", () => {
 
   test("should be mobile responsive", async ({ page }) => {
     await page.setViewportSize({ width: 375, height: 667 });
-    await presetCookieConsent(page, "/time-tracker-privacy.html");
+    await page.goto("/time-tracker-privacy.html");
+
+    // Handle cookie consent
+    await handleCookieConsent(page);
 
     // Content should be visible
     const section = page.locator("section").first();
@@ -191,8 +197,10 @@ test.describe("Time Tracker Privacy Policy Page", () => {
 
 test.describe("Projects Section Integration", () => {
   test("should navigate from main page to privacy policy", async ({ page }) => {
-    // Pre-set cookie consent to prevent dialog from appearing
-    await presetCookieConsent(page, "/");
+    await page.goto("/");
+
+    // Handle cookie consent
+    await handleCookieConsent(page);
 
     // Scroll to Projects section
     await page.locator("#projects-title").scrollIntoViewIfNeeded();
@@ -211,8 +219,10 @@ test.describe("Projects Section Integration", () => {
   });
 
   test("should navigate from footer to privacy policy", async ({ page }) => {
-    // Pre-set cookie consent to prevent dialog from appearing
-    await presetCookieConsent(page, "/");
+    await page.goto("/");
+
+    // Handle cookie consent
+    await handleCookieConsent(page);
 
     // Find footer link
     const footerPrivacyLink = page
@@ -229,8 +239,10 @@ test.describe("Projects Section Integration", () => {
   });
 
   test("should show Projects in navigation menu", async ({ page }) => {
-    // Pre-set cookie consent to prevent dialog from appearing
-    await presetCookieConsent(page, "/");
+    await page.goto("/");
+
+    // Handle cookie consent
+    await handleCookieConsent(page);
 
     const projectsLink = page.locator('a[href="#projects-title"]');
 
@@ -251,8 +263,10 @@ test.describe("Projects Section Integration", () => {
   });
 
   test("should display Time Tracker project card", async ({ page }) => {
-    // Pre-set cookie consent to prevent dialog from appearing
-    await presetCookieConsent(page, "/");
+    await page.goto("/");
+
+    // Handle cookie consent
+    await handleCookieConsent(page);
 
     // Scroll to projects
     await page.locator("#projects-title").scrollIntoViewIfNeeded();
@@ -269,8 +283,10 @@ test.describe("Projects Section Integration", () => {
 
 test.describe("Cross-Page Theme Consistency", () => {
   test("should maintain consistent styling across pages", async ({ page }) => {
-    // Check main page
-    await presetCookieConsent(page, "/");
+    await page.goto("/");
+
+    // Handle cookie consent
+    await handleCookieConsent(page);
 
     const mainNavBg = await page
       .locator("nav")
@@ -288,8 +304,10 @@ test.describe("Cross-Page Theme Consistency", () => {
   });
 
   test("should load shared theme.js on both pages", async ({ page }) => {
-    // Test on main page
-    await presetCookieConsent(page, "/");
+    await page.goto("/");
+
+    // Handle cookie consent
+    await handleCookieConsent(page);
 
     let toggleThemeExists = await page.evaluate(
       () => typeof toggleTheme === "function"
@@ -308,7 +326,10 @@ test.describe("Cross-Page Theme Consistency", () => {
 
 test.describe("Privacy Policy Content Quality", () => {
   test("should have GDPR compliance sections", async ({ page }) => {
-    await presetCookieConsent(page, "/time-tracker-privacy.html");
+    await page.goto("/time-tracker-privacy.html");
+
+    // Handle cookie consent
+    await handleCookieConsent(page);
 
     const content = await page.textContent("body");
 
@@ -319,7 +340,10 @@ test.describe("Privacy Policy Content Quality", () => {
   });
 
   test("should have data retention information", async ({ page }) => {
-    await presetCookieConsent(page, "/time-tracker-privacy.html");
+    await page.goto("/time-tracker-privacy.html");
+
+    // Handle cookie consent
+    await handleCookieConsent(page);
 
     // Check for data retention section
     await expect(
@@ -333,7 +357,10 @@ test.describe("Privacy Policy Content Quality", () => {
   });
 
   test("should explain data collection clearly", async ({ page }) => {
-    await presetCookieConsent(page, "/time-tracker-privacy.html");
+    await page.goto("/time-tracker-privacy.html");
+
+    // Handle cookie consent
+    await handleCookieConsent(page);
 
     // Should have numbered or structured information collection
     await expect(
